project3/pyproject.toml
[project]
name = "project3"
version = "0.1.0"
description = "Add your description here"
readme = "README.md"
requires-python = ">=3.13"
dependencies = [
    "google-generativeai>=0.8.5",
    "pandas>=2.3.0",
    "python-dotenv>=1.1.1",
    "streamlit>=1.46.1",
]
